Implement Token and validate methods whit login.

+++ src/main/java/com/procces/business/app/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.procces.business.app.controllers;
 
 import com.procces.business.app.models.User;
 import com.procces.business.app.services.UserService;
+import com.procces.business.app.utils.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,8 +21,14 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private JWTUtil jwtUtil;
+
     @GetMapping(value = "/user/{id}")
-    public ResponseEntity getUser(@PathVariable Long id) {
+    public ResponseEntity getUser(@PathVariable Long id, @RequestHeader(value = "Authorization") String token) {
+        if (jwtUtil.getKey(token) == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token invalid");
+        }
         return userService.getUser(id);
     }
 
@@ -31,32 +38,47 @@ public class UserController {
     }
 
     @GetMapping(value = "/users")
-    public ResponseEntity listUser() {
+    public ResponseEntity listUser(@RequestHeader(value = "Authorization") String token) {
+        if (jwtUtil.getKey(token) == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token invalid");
+        }
         return userService.listUser();
     }
 
     @GetMapping(value = "/user/{name}/{lastName}")
-    public ResponseEntity getByNameAndLastName(@PathVariable String name, @PathVariable String lastName) {
+    public ResponseEntity getByNameAndLastName(@PathVariable String name, @PathVariable String lastName, @RequestHeader(value = "Authorization") String token) {
+        if (jwtUtil.getKey(token) == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token invalid");
+        }
         return userService.getByNameAndLastName(name, lastName);
     }
 
     @GetMapping(value = "/user/lastname/{lastName}")
-    public ResponseEntity getByLastName(@PathVariable String lastName) {
+    public ResponseEntity getByLastName(@PathVariable String lastName, @RequestHeader(value = "Authorization") String token) {
+        if (jwtUtil.getKey(token) == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token invalid");
+        }
         return userService.getByLastName(lastName);
     }
 
     @GetMapping(value = "/user/name/{name}")
-    public ResponseEntity getByName(@PathVariable String name) {
+    public ResponseEntity getByName(@PathVariable String name, @RequestHeader(value = "Authorization") String token) {
         return userService.getByName(name);
     }
 
     @PutMapping(value = "/user/{id}")
-    public ResponseEntity updateUser(@Valid @PathVariable Long id, @RequestBody User user) {
+    public ResponseEntity updateUser(@Valid @PathVariable Long id, @RequestBody User user, @RequestHeader(value = "Authorization") String token) {
+        if (jwtUtil.getKey(token) == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token invalid");
+        }
         return userService.updateUser(id, user);
     }
 
     @DeleteMapping(value = "/user/{id}")
-    public ResponseEntity deleteUser(@PathVariable Long id) {
+    public ResponseEntity deleteUser(@PathVariable Long id, @RequestHeader(value = "Authorization") String token) {
+        if (jwtUtil.getKey(token) == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token invalid");
+        }
         return userService.deleteUser(id);
     }
